refactor: change folders inside infrastructure folder to improve DDD organization based on bounded contexts
