+++ src/infrastructure/order/repository/sequelize/order-item.model.ts
@@ -1,6 +1,6 @@
 import {Model, BelongsTo, Column, ForeignKey, PrimaryKey, Table} from "sequelize-typescript";
 import {DataTypes} from "sequelize";
-import ProductModel from "./product.model";
+import ProductModel from "../../../product/repository/sequelize/product.model";
 import OrderModel from "./order.model";
 
 @Table({tableName: "order-items", timestamps: false})
