+++ src/infrastructure/order/repository/sequelize/order.repository.ts
@@ -1,6 +1,6 @@
 import Order from "@domain/checkout/entity/order";
-import OrderModel from "@infra/db/sequelize/model/order.model";
-import OrderItemModel from "@infra/db/sequelize/model/order-item.model";
+import OrderModel from "@infra/order/repository/sequelize/order.model";
+import OrderItemModel from "@infra/order/repository/sequelize/order-item.model";
 import IOrderRepository from "@domain/checkout/repository/order-repository.interface";
 import {Promise} from "ts-toolbelt/out/Any/Promise";
 import OrderItem from "@domain/checkout/entity/order_item";
